pyproject.toml
[project]
name = "gnn-recsys"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "pyarrow>=14.0.1",
    "numpy",
    "scikit-learn",
    "pydantic",
    "polars",
    "duckdb",
    "matplotlib",
    "networkx",
    "seaborn",
    "pyyaml",
    "gensim",
    "ogb",
    "dgl==2.1.0",
    "torch==2.2.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
dgl = {url = "https://data.dgl.ai/wheels/cu118/dgl-2.1.0%2Bcu118-cp310-cp310-manylinux1_x86_64.whl"}

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv]
upgrade = false
dev-dependencies = ["pre-commit", "ipykernel", "ipywidgets"]

[tool.ruff]
lint.ignore = ["E402", "F841"]
lint.select = ["E", "F", "I", "W"]
lint.fixable = ["I", "F401"]
line-length = 120
include = ["*.py"]
exclude = [".env", ".venv", "venv", "notebooks", "ref"]
output-format = "full"
src = ["src"]

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-local-folder = ["src", "scripts"]
no-lines-before = ["local-folder"]

[tool.mypy]
disallow_untyped_defs = true
disallow_any_unimported = false # this is needed to pass for torch and dgl
check_untyped_defs = true
show_error_codes = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
disable_error_code = ["dict-item"]
exclude = ["scripts/.*", "data/.*", "notebooks/.*", "ref/.*"]


[[tool.mypy.overrides]]
module = [
    "dgl.*",
    "torch.*",
    "torch.nn.*",
    "torch.optim.*",
    "torch.utils.checkpoint.*",
    "torch.utils.data.*",
    "torch.utils.dlpack.*",
    "torch.utils.tensorboard.*",
    "torch.utils.tensorboard.summary.*",
    "torch.utils.tensorboard.writer.*",
]
ignore_errors = true
ignore_missing_imports = true
follow_imports = "silent"
disable_error_code = ["no-any-unimported"] 